+++ src/lib.rs
@@ -2,11 +2,13 @@
 #![doc = include_str!("../README.md")]
 
 pub mod light;
+pub mod occluder;
 pub mod plugin;
 mod render;
 
 /// A module which exports commonly used dependencies.
 pub mod prelude {
     pub use crate::light::{AmbientLight2d, PointLight2d, PointLight2dBundle};
+    pub use crate::occluder::{LightOccluder2d, LightOccluder2dBundle, LightOccluder2dShape};
     pub use crate::plugin::Light2dPlugin;
 }