+++ src/render/light_map/pipeline.rs
@@ -1,0 +1,77 @@
+use bevy::core_pipeline::fullscreen_vertex_shader::fullscreen_shader_vertex_state;
+use bevy::ecs::system::Resource;
+use bevy::ecs::world::{FromWorld, World};
+use bevy::render::render_resource::binding_types::{sampler, texture_2d, uniform_buffer};
+use bevy::render::render_resource::{
+    BindGroupLayout, BindGroupLayoutEntries, CachedRenderPipelineId, ColorTargetState, ColorWrites,
+    FragmentState, GpuArrayBuffer, MultisampleState, PipelineCache, PrimitiveState,
+    RenderPipelineDescriptor, Sampler, SamplerBindingType, SamplerDescriptor, ShaderStages,
+    TextureFormat, TextureSampleType,
+};
+use bevy::render::renderer::RenderDevice;
+use bevy::render::view::ViewUniform;
+
+use crate::render::extract::{ExtractedAmbientLight2d, ExtractedPointLight2d};
+
+use super::LIGHT_MAP_SHADER;
+
+const LIGHT_MAP_BIND_GROUP_LAYOUT: &str = "light_map_group_layout";
+const LIGHT_MAP_PIPELINE: &str = "light_map_pipeline";
+
+#[derive(Resource)]
+pub struct LightMapPipeline {
+    pub layout: BindGroupLayout,
+    pub sdf_sampler: Sampler,
+    pub pipeline_id: CachedRenderPipelineId,
+}
+
+impl FromWorld for LightMapPipeline {
+    fn from_world(world: &mut World) -> Self {
+        let render_device = world.resource::<RenderDevice>();
+
+        let layout = render_device.create_bind_group_layout(
+            LIGHT_MAP_BIND_GROUP_LAYOUT,
+            &BindGroupLayoutEntries::sequential(
+                ShaderStages::FRAGMENT,
+                (
+                    uniform_buffer::<ViewUniform>(true),
+                    uniform_buffer::<ExtractedAmbientLight2d>(true),
+                    GpuArrayBuffer::<ExtractedPointLight2d>::binding_layout(render_device),
+                    texture_2d(TextureSampleType::Float { filterable: true }),
+                    sampler(SamplerBindingType::Filtering),
+                ),
+            ),
+        );
+
+        let sdf_sampler = render_device.create_sampler(&SamplerDescriptor::default());
+
+        let pipeline_id =
+            world
+                .resource_mut::<PipelineCache>()
+                .queue_render_pipeline(RenderPipelineDescriptor {
+                    label: Some(LIGHT_MAP_PIPELINE.into()),
+                    layout: vec![layout.clone()],
+                    vertex: fullscreen_shader_vertex_state(),
+                    fragment: Some(FragmentState {
+                        shader: LIGHT_MAP_SHADER,
+                        shader_defs: vec![],
+                        entry_point: "fragment".into(),
+                        targets: vec![Some(ColorTargetState {
+                            format: TextureFormat::Rgba16Float,
+                            blend: None,
+                            write_mask: ColorWrites::ALL,
+                        })],
+                    }),
+                    primitive: PrimitiveState::default(),
+                    depth_stencil: None,
+                    multisample: MultisampleState::default(),
+                    push_constant_ranges: vec![],
+                });
+
+        Self {
+            layout,
+            sdf_sampler,
+            pipeline_id,
+        }
+    }
+}