+++ src/render/lighting/mod.rs
@@ -12,8 +12,8 @@ use bevy::{
 };
 
 pub use node::LightingNode;
-pub use pipeline::LightingPipeline;
-pub use prepare::prepare_lighting_pipelines;
+pub use pipeline::*;
+pub use prepare::*;
 
 pub const LIGHTING_SHADER: Handle<Shader> =
     Handle::weak_from_u128(111120241052143214281687226997564407636);